requirements.txt
discord.py
python-dotenv
yt-dlp
pynacl
